adding in logic to ask user to verify if they want to delete the project.

+++ ppmtool-react-client/src/actions/projectActions.js
@@ -41,9 +41,15 @@ export const getProject = (id, history) => async dispatch => {
 };
 
 export const deleteProject = id => async dispatch => {
-  await axios.delete(`${Constants.PROJECT_API_URL}/${id}`);
-  dispatch({
-    type: DELETE_PROJECT,
-    payload: id
-  });
+  if (
+    window.confirm(
+      "Are you sure? This will delete the project and all the data related to it"
+    )
+  ) {
+    await axios.delete(`${Constants.PROJECT_API_URL}/${id}`);
+    dispatch({
+      type: DELETE_PROJECT,
+      payload: id
+    });
+  }
 };
